revised hasTextMonth parsing and 2digitYear parsing
I re-wrote the parsing to determine if a token has a text month.  This version is now more flexible and can handle extraneous spaces and punctuation that my new temporal parsing method does not remove.  I also updated the 2-digit year method to return proper relative spans in case there are multiple substrings of the year in the full phrase.  Like "910203 02/03/91" previously would always return the span of the first "91" it found rather than the second one.