updated filtering of temporal phrases list
Included code to determine if two consecutive tokens actually had a newline inbetween them.  Turns out the end span of the first plus 1 should equal the start span of the next if there are no new lines.  New lines increment the spans so they don't match when just adding 1.  This new code checks for this so that temporal tokens on multiple lines are not mushed together is one long temporal phrase.  I also corrected the parsing so that lone numeric tokens that were also flagged as temporal were not removed from my parsing.