Added new method getRefIdx()
Added a new method in utils that takes the reference list, a start span, and an end span.  It compares the start and end and returns the index of the ref_list that overlaps with it.  This helps us take a temporal phrase and identify where it came from in the reference list.  I then included this method in the buildPeriodInterval() code.