Added additional Before words to parse
Added the words "earlier" and "until" to the list of Before words for parsing.  Also updated the buildBeforeAfter() methods to set the Interval-Type to "Link" instead of "DocTime" (which is still the default if no paramerter is passed).  This is naive and now getting all the DocTime interval types wrong.  Need to this of an intelleigent way to parse this information out.  Also, I updated the runit.sh script to ignore all Before and After Intervals as these are the links to events that we are not parsing out yet.