Fixed bug in hasTextMonth, included identification of modifier words in build subintervals
Lots a changes in this commit.  First I included checking for a modifier word like "next", "this", or "last" in the same temporal phrase as a Month without a Year entity.  If one of these exists it skips it and only tags that months without a Year and without a Modifier.  Second, I re-wrote the hasTextMonth() method so that it identifies months at the end of sentences.  It was missing things like "December." because I wasn't removing the period.  The way I re-wrote this should be used as a template to re-write similar methods because this is a more robust way to do it.