major bug fix for getTemporalPhrases()!  now processing last array element
When I wrote this method I didn't understand how python indexes and how the range function worked exactly.  Turns out I was skipping the last element of the reference token array!  Important time data was missing!  The range function is range(inclusive, exclusive) where the second bound is not actually included.  Thus the last element of the array is always 1 less than the length of the array.  This is fixed now and we are getting additional elements!