updated time zones to return correct spans
With the re.search() method you have to specify the group number you want the spans for.  In this case it is group 1.  I simplified the code to remove the complex parsing.  Note that I removed the digit specification because it was zero or more, which is the same as not looking for digits at all.  We DO get false positives at this step, but in the buildSubIntervals method I added a phrase to remove a tz entity if there was no corresponding hour to link it to in the temporal phrase.