Added sentence boundry parsing to the getWhitespaceTokens method
Primary changes were to add sentence boundry parsing to the getWhitespaceTokens utils method.  This method now returns a boolean array with one element per reference token where it is set to 1 if that token is either the last in a sentence or the last on a line.  I also updated the getTemporalPhrase method to utilize this new property of the reference tokens.  If it finds the token it is on is at the end of a sentence or line then it ends the temporal phrase.  Currently, the ML methods do not utilize this new property when generating feature vectors.